fix: explicitly cast double value in status endpoint
## Problem
- Server status endpoint still returning 'Bad cast exception' after previous fix
- Issue was with double value 'share' in thread pool status not being properly handled

## Root Cause
- Poco::JSON::Object::set() with double values can cause bad cast exceptions
- The 'share' field in ThreadPoolManager::TypeStatus is a double type
- Need explicit static_cast<double>() to ensure proper type handling

## Solution
- Added explicit cast: type_obj->set("share", static_cast<double>(kv.second.share))
- This ensures the double value is properly handled by Poco::JSON

## Files Modified
- src/core/webserver/web_handlers_server_status.cpp: Added explicit double cast

## Testing
- Build successful
- Server needs restart to apply fix

## Impact
- Should resolve the remaining 'Bad cast exception' in status endpoint
- Frontend should be able to fetch server status correctly

+++ src/core/webserver/web_handlers_server_status.cpp
@@ -123,7 +123,7 @@ namespace MediaDedup
                 for (const auto &kv : tpm_status.perType)
                 {
                     Poco::JSON::Object::Ptr type_obj = new Poco::JSON::Object();
-                    type_obj->set("share", kv.second.share);
+                    type_obj->set("share", static_cast<double>(kv.second.share));
                     type_obj->set("running", static_cast<int>(kv.second.running));
                     type_obj->set("queued", static_cast<int>(kv.second.queued));
                     per_type_obj->set(kv.first, type_obj);
